Current Progress
Includes my statistics module which fits the needs of this project, some functions for loan costs, a "city" which represents the locations of houses in the simulation (which will play in to a future feature) and the work in progress order book which will match buyers to sellers. Soon going to be refactoring things to include classes instead of the ghetto way I'm doing houses now.